inference/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
awslambdaric==2.2.1
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.2
fastapi==0.114.1
fastapi-cli==0.0.5
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.0
groq==0.11.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.8
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-anthropic==0.2.3
langchain-community==0.3.0
langchain-core==0.3.13
langchain-groq==0.2.0
langchain-openai==0.2.4
langchain-text-splitters==0.3.0
langcodes==3.4.0
langsmith==0.1.137
language_data==1.2.0
mangum==0.19.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.52.2
orjson==3.10.7
packaging==24.1
pinecone==5.1.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pip-autoremove==0.10.0
preshed==3.0.9
prisma==0.15.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
shellingham==1.5.4
simplejson==3.19.3
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.34
srsly==2.4.8
starlette==0.38.5
StrEnum==0.4.15
tenacity==8.5.0
thinc==8.2.5
tiktoken==0.8.0
tokenizers==0.20.1
tomlkit==0.13.2
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
voyageai==0.2.4
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websockets==13.0.1
wrapt==1.16.0
yarl==1.11.1
